+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.3.2 -
 * Refactor calls to Invoke-RestMethod to Invoke-MIARequest
 * Consolidate .Format.ps1xml file
+* Add `*-MIATaskScheduler` functions
 ## 0.3.1 - March 2022
 * Fixed issue in Wait-MIATask on Windows PowerShell 5.1
 * Add Get-MIACustomScript command