+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.3.2 -
+## 0.3.2 - June 2022
 * Refactor calls to Invoke-RestMethod to Invoke-MIARequest
 * Consolidate .Format.ps1xml file
 * Add `*-MIATaskScheduler` functions